Trait sui_data_ingestion_core::Worker

pub trait Worker: Send + Sync {
    // Required method
    fn process_checkpoint<'life0, 'async_trait>(
        &'life0 self,
        checkpoint: CheckpointData
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn save_progress<'life0, 'async_trait>(
        &'life0 self,
        sequence_number: CheckpointSequenceNumber
    ) -> Pin<Box<dyn Future<Output = Option<CheckpointSequenceNumber>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn preprocess_hook(&self, _: CheckpointData) -> Result<()> { ... }
}

Required Methods

fn process_checkpoint<'life0, 'async_trait>( &'life0 self, checkpoint: CheckpointData ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Provided Methods

fn save_progress<'life0, 'async_trait>( &'life0 self, sequence_number: CheckpointSequenceNumber ) -> Pin<Box<dyn Future<Output = Option<CheckpointSequenceNumber>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Optional method. Allows controlling when workflow progress is updated in the progress store. For instance, some pipelines may benefit from aggregating checkpoints, thus skipping the saving of updates for intermediate checkpoints. The default implementation is to update the progress store for every processed checkpoint.

fn preprocess_hook(&self, _: CheckpointData) -> Result<()>