sui_data_ingestion_core
Trait Worker

pub trait Worker: Send + Sync {
    type Result: Send + Sync + Clone;

    // Provided methods
    fn process_checkpoint_arc<'life0, 'life1, 'async_trait>(
        &'life0 self,
        checkpoint: &'life1 Arc<CheckpointData>,
    ) -> Pin<Box<dyn Future<Output = Result<Self::Result>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn process_checkpoint<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _checkpoint_data: &'life1 CheckpointData,
    ) -> Pin<Box<dyn Future<Output = Result<Self::Result>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn preprocess_hook(&self, _: &CheckpointData) -> Result<()> { ... }
}

Required Associated Types

type Result: Send + Sync + Clone

Provided Methods

fn process_checkpoint_arc<'life0, 'life1, 'async_trait>( &'life0 self, checkpoint: &'life1 Arc<CheckpointData>, ) -> Pin<Box<dyn Future<Output = Result<Self::Result>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn process_checkpoint<'life0, 'life1, 'async_trait>( &'life0 self, _checkpoint_data: &'life1 CheckpointData, ) -> Pin<Box<dyn Future<Output = Result<Self::Result>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

There is no need to implement this if you implement process_checkpoint_arc. The WorkerPool will only call process_checkpoint_arc. This method was left in place for backwards compatibiity.

fn preprocess_hook(&self, _: &CheckpointData) -> Result<()>